Split DAUChart into its own file
